requirements.txt
fastapi==0.65.2
uvicorn==0.11.8
motor==2.2.0
python-decouple==3.3
pydantic[email]
PyJWT==1.7.1
passlib[bcrypt]
requests==2.26.0